pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "tkan"
version = "0.5.0"
description = "Temporal KAN model"
authors = ["Rémi Genet", "Hugo Inzirillo"]
readme = "README.md"
packages = [{include = "tkan"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
jax = "^0.4.13"
jaxlib = "^0.4.13"
optax = "^0.1.7"
jax2onnx = "^0.1"
pyyaml = "^6.0"
pandas = "^2.0"
pyarrow = "^14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
filterwarnings = [
    "ignore:Can't initialize NVML:UserWarning",
    "ignore:jax.xla_computation is deprecated:DeprecationWarning",
    "ignore::DeprecationWarning:jax._src.dtypes",
]
